ARM64: configs: gauguin: Disable sysrq keys over serial
If the UART console is enabled in the bootloader, this could cause false
triggers, especially if the other side's UART interface has a loose TX
pin (which I've encountered).

Note that sysrq is still enabled because Android init uses it when
shutting down, and it doesn't hurt to keep it enabled for debugging.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Change-Id: Ida5af77e4fe7d5ad9f7bfd3c10e3ae10ed281042
Signed-off-by: hridaya <hridaya@pixelos.net>